+++ src/actions/approve.js
@@ -1,0 +1,3 @@
+export const approve = (selectedRows) => {
+    alert(`Approved ${selectedRows.map(row => row.name).join(', ')}`);
+};
