+++ src/actions/copy.js
@@ -1,0 +1,3 @@
+export const copy = (selectedRows) => {
+    alert(`Copied ${selectedRows.map(row => row.name).join(', ')}`);
+};
