+++ src/actions/viewAttachments.js
@@ -1,0 +1,3 @@
+export const viewAttachments = (selectedRows) => {
+    alert(`Viewing attachments for ${selectedRows.map(row => row.name).join(', ')}`);
+};
