+++ src/components/App.jsx
@@ -1,5 +1,6 @@
 import { DynamicTable } from './DynamicTable';
 import { Check, AttachFile, FileCopy } from '@mui/icons-material';
+import { approve, viewAttachments, copy } from '../actions';
 
 const columns = [
     { field: 'id', headerName: 'ID' },
@@ -31,19 +32,19 @@ const tools = [
         name: 'Approve',
         icon: <Check />,
         tooltip: 'Approve selected rows',
-        action: (selectedRows) => alert(`Approved ${selectedRows.map(row => row.name).join(', ')}`),
+        action: approve,
     },
     {
         name: 'View Attachments',
         icon: <AttachFile />,
         tooltip: 'View attachments for selected rows',
-        action: (selectedRows) => alert(`Viewing attachments for ${selectedRows.map(row => row.name).join(', ')}`),
+        action: viewAttachments,
     },
     {
         name: 'Copy',
         icon: <FileCopy />,
         tooltip: 'Copy selected rows',
-        action: (selectedRows) => alert(`Copied ${selectedRows.map(row => row.name).join(', ')}`),
+        action: copy,
     },
 ];
