+++ src/components/DynamicTable.jsx
